Issue: rapidjson maintains its own hierarchical structure and can reallocate and copy lists as needed to expand the number of children of a node in real time. the Property tree keeps pointers into these structures and those pointers can orphaned after a reallocate/copy.
So any operation that adds a child to a node (which could potentially trigger a reallocate/copy)
incrementes a counter.

Any property element access checks the value of the saved counter against the current counter
and if the value is old, will re-find it's actual pointer in the json structure before looking
up the requested value and returning it.

This isn't great stuff, but typically once the tree is built it's not being added to so this
doesn't actually cause that much extra work.

However, I discovered the update the tree from a json string code doesn't fully work as expected
and can lead to a memory leak if you send a multi-level json tree.  Ugh.